requirements.txt
Click==7.0
Flask==1.1.1
Flask-SocketIO==4.2.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psutil==5.6.3
pypiwin32==223
python-engineio==3.9.3
python-socketio==4.3.1
pywin32==225
pywinservicemanager==1.0.2
six==1.12.0
Werkzeug==0.15.6
